Tests: Move default test files to test/main
Fix #781
