+++ test/headless.html
@@ -5,8 +5,8 @@
 	<title>QUnit Headless Test Suite</title>
 	<link rel="stylesheet" href="../dist/qunit.css">
 	<script src="../dist/qunit.js"></script>
-	<script src="test.js"></script>
-	<script src="deepEqual.js"></script>
+	<script src="main/test.js"></script>
+	<script src="main/deepEqual.js"></script>
 	<script>
 		(function () {
 			function createCallback(logType) {
