+++ test/index.html
@@ -5,15 +5,15 @@
 	<title>QUnit Main Test Suite</title>
 	<link rel="stylesheet" href="../dist/qunit.css">
 	<script src="../dist/qunit.js"></script>
-	<script src="test.js"></script>
+	<script src="main/test.js"></script>
 	<script src="reporter-html.js"></script>
-	<script src="assert.js"></script>
-	<script src="async.js"></script>
-	<script src="promise.js"></script>
-	<script src="dump.js"></script>
-	<script src="modules.js"></script>
-	<script src="deepEqual.js"></script>
-	<script src="globals.js"></script>
+	<script src="main/assert.js"></script>
+	<script src="main/async.js"></script>
+	<script src="main/promise.js"></script>
+	<script src="main/dump.js"></script>
+	<script src="main/modules.js"></script>
+	<script src="main/deepEqual.js"></script>
+	<script src="main/globals.js"></script>
 	<script src="swarminject.js"></script>
 </head>
 <body>
