+++ test/main/globals.js
