+++ test/main/promise.js
